ACL_TensorFlow/contrib/cv/PRSR_ID2111_for_ACL/requirements.txt
numpy~=1.18.5
pillow~=7.2.0
h5py~=2.10.0
tensorflow~=1.15